+++ libc/src/__support/StringUtil/tables/minimal_platform_signals.h
@@ -13,7 +13,7 @@
 
 namespace __llvm_libc {
 
-inline constexpr auto PLATFORM_SIGNALS = STDC_SIGNALS;
+LIBC_INLINE_VAR constexpr auto PLATFORM_SIGNALS = STDC_SIGNALS;
 
 } // namespace __llvm_libc